// frontend/generated/openapi-client/types.gen.ts
export type Account = {
    id: string;
    userId: string;
    type: string;
    provider: string;
    providerAccountId: string;
    refresh_token?: string;
    access_token?: string;
    expires_at?: number;
    token_type?: string;
    scope?: string;
    id_token?: string;
    session_state?: string;
    user: User;
};

export type Session = {
    id: string;
    sessionToken: string;
    userId: string;
    expires: string;
    user: User;
};

export type CourseCategory = {
    id: string;
    name: string;
    slug: string;
    description?: string;
    createdAt: string;
    updatedAt: string;
    courses: Array<Course>;
};

export type CourseEnrollment = {
    id: string;
    userId: string;
    courseId: string;
    enrolledAt: string;
    createdAt: string;
    updatedAt: string;
    course: Course;
    user: User;
};

export type CourseReview = {
    id: string;
    content: string;
    rating: number;
    userId: string;
    courseId: string;
    instructorReply?: string;
    createdAt: string;
    updatedAt: string;
    course: Course;
    user: User;
};

export type CourseComment = {
    id: string;
    content: string;
    userId: string;
    questionId: string;
    createdAt: string;
    updatedAt: string;
    question: CourseQuestion;
    user: User;
};

export type CourseQuestion = {
    id: string;
    title: string;
    content: string;
    userId: string;
    courseId: string;
    createdAt: string;
    updatedAt: string;
    comments: Array<CourseComment>;
    course: Course;
    user: User;
};

export type Course = {
    id: string;
    slug: string;
    title: string;
    shortDescription?: string;
    description?: string;
    thumbnailUrl?: string;
    price: number;
    discountPrice?: number;
    level: string;
    status: string;
    instructorId: string;
    isPublished: boolean;
    createdAt: string;
    updatedAt: string;
    sections: Array<Section>;
    lectures: Array<Lecture>;
    categories: Array<CourseCategory>;
    enrollments: Array<CourseEnrollment>;
    reviews: Array<CourseReview>;
    questions: Array<CourseQuestion>;
    instructor: User;
};

export type User = {
    id: string;
    name?: string;
    email?: string;
    emailVerified?: string;
    hashedPassword?: string;
    image?: string;
    accounts: Array<Account>;
    sessions: Array<Session>;
    courses: Array<Course>;
    courseEnrollments: Array<CourseEnrollment>;
    courseReviews: Array<CourseReview>;
    courseQuestions: Array<CourseQuestion>;
    courseComments: Array<CourseComment>;
    lectureActivities: Array<LectureActivity>;
};

export type LectureActivity = {
    id: string;
    userId: string;
    lectureId: string;
    progress: number;
    isCompleted: boolean;
    lastWatchedAt: string;
    lecture: Lecture;
    user: User;
};

export type Lecture = {
    id: string;
    title: string;
    description?: string;
    order: number;
    duration?: number;
    isPreview: boolean;
    sectionId: string;
    courseId: string;
    videoStorageInfo?: {
        [key: string]: unknown;
    };
    createdAt: string;
    updatedAt: string;
    activities: Array<LectureActivity>;
    section: Section;
    course: Course;
};

export type Section = {
    id: string;
    title: string;
    description?: string;
    order: number;
    courseId: string;
    createdAt: string;
    updatedAt: string;
    lectures: Array<Lecture>;
    course: Course;
};